Add integer cast

+++ src/main/java/de/chojo/repbot/dao/provider/Guilds.java
@@ -70,7 +70,7 @@ public class Guilds {
     private Integer pages(int pageSize) {
         return query("""
                        SELECT
-                           CEIL(COUNT(1)::numeric / ?) AS count
+                           CEIL(COUNT(1)::numeric / ?)::INTEGER AS count
                        FROM
                            guilds
                        """)
